Reimplement DocsHeading component
